+++ src/jira/createJiraClient.ts
@@ -1,9 +1,20 @@
-import JiraClient from "./JiraClient";
+import { IntegrationLogger } from '@jupiterone/integration-sdk-core';
+import JiraClient from './JiraClient';
 
-export default function createJiraClient(config: any) {
-  return new JiraClient({
-    host: config.jiraHost,
-    username: config.jiraUsername,
-    password: config.jiraPassword,
-  });
+export default function createJiraClient(
+  config: {
+    jiraHost: string;
+    jiraUsername: string;
+    jiraPassword: string;
+  },
+  logger: IntegrationLogger,
+) {
+  return new JiraClient(
+    {
+      host: config.jiraHost,
+      username: config.jiraUsername,
+      password: config.jiraPassword,
+    },
+    logger,
+  );
 }