+++ src/validateInvocation.test.ts
@@ -20,14 +20,13 @@ it('requires valid config', async () => {
   );
 });
 
-
-test("should throw exception if jiraHost has invalid chars", async () => {
+test('should throw exception if jiraHost has invalid chars', async () => {
   const context = {
     instance: {
       config: {
-        jiraHost: "test.com?somequeryparms",
-        jiraUsername: "testLogin",
-        jiraPassword: "testPassword",
+        jiraHost: 'test.com?somequeryparms',
+        jiraUsername: 'testLogin',
+        jiraPassword: 'testPassword',
       },
     },
   };
@@ -39,13 +38,13 @@ test("should throw exception if jiraHost has invalid chars", async () => {
   }
 });
 
-test("should throw auth error not instanceConfigError if jiraHost has a subdir", async () => {
+test('should throw auth error not instanceConfigError if jiraHost has a subdir', async () => {
   const context = {
     instance: {
       config: {
-        jiraHost: "test.com/subdir",
-        jiraUsername: "testLogin",
-        jiraPassword: "testPassword",
+        jiraHost: 'fake-hostname.atlassian.net/subdir',
+        jiraUsername: 'testLogin',
+        jiraPassword: 'testPassword',
       },
     },
   };